arcticapi/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
jedi==0.16.0
parso==0.6.2
prompt-toolkit==3.0.3
ptpython==3.0.1
Pygments==2.5.2
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.44.0
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==5.0.1
